Add block template with abstract blocks and templates (#39630)
* Make block template abstract and protected add block methods

* Create block abstraction and generic block

* Remove add_block from container interface

* Update tests for generic and custom blocks

* Add tests around custom block templates

* Fix up lint errors

* Fix errant comment for add_block
